Refactor to use only composite task (WIP)
