+++ cmd/ably-boomer/main.go
@@ -39,62 +39,6 @@ func run(taskFn func(), c *cli.Context) error {
 	return nil
 }
 
-func runFanOut(c *cli.Context) error {
-	apiKey := c.String(apiKeyFlag.Name)
-	env := c.String(envFlag.Name)
-	channelName := c.String(channelNameFlag.Name)
-	taskFn := ably.CurryFanOutTask(ably.FanOutConf{
-		Logger:      log,
-		APIKey:      apiKey,
-		Env:         env,
-		ChannelName: channelName,
-	})
-
-	return run(taskFn, c)
-}
-
-func runPersonal(c *cli.Context) error {
-	apiKey := c.String(apiKeyFlag.Name)
-	env := c.String(envFlag.Name)
-	publishInterval := c.Int(publishIntervalFlag.Name)
-	numSubscriptions := c.Int(numSubscriptionsFlag.Name)
-	msgDataLength := c.Int(msgDataLengthFlag.Name)
-	sseSubscriber := c.Bool(sseSubscriberFlag.Name)
-	taskFn := ably.CurryPersonalTask(ably.PersonalConf{
-		Logger:           log,
-		APIKey:           apiKey,
-		SSESubscriber:    sseSubscriber,
-		Env:              env,
-		PublishInterval:  publishInterval,
-		NumSubscriptions: numSubscriptions,
-		MsgDataLength:    msgDataLength,
-	})
-
-	return run(taskFn, c)
-}
-
-func runSharded(c *cli.Context) error {
-	apiKey := c.String(apiKeyFlag.Name)
-	env := c.String(envFlag.Name)
-	numChannels := c.Int(numChannelsFlag.Name)
-	publishInterval := c.Int(publishIntervalFlag.Name)
-	msgDataLength := c.Int(msgDataLengthFlag.Name)
-	numSubscriptions := c.Int(numSubscriptionsFlag.Name)
-	publisher := c.Bool(publisherFlag.Name)
-	taskFn := ably.CurryShardedTask(ably.ShardedConf{
-		Logger:           log,
-		APIKey:           apiKey,
-		Env:              env,
-		NumChannels:      numChannels,
-		PublishInterval:  publishInterval,
-		MsgDataLength:    msgDataLength,
-		NumSubscriptions: numSubscriptions,
-		Publisher:        publisher,
-	})
-
-	return run(taskFn, c)
-}
-
 func runComposite(c *cli.Context) error {
 	apiKey := c.String(apiKeyFlag.Name)
 	env := c.String(envFlag.Name)
@@ -122,46 +66,6 @@ func main() {
 
 	app := &cli.App{
 		Commands: []*cli.Command{
-			{
-				Name:    "fanout",
-				Aliases: []string{"f"},
-				Usage:   "run a fanout task",
-				Action:  runFanOut,
-				Flags: []cli.Flag{
-					apiKeyFlag,
-					envFlag,
-					channelNameFlag,
-				},
-			},
-			{
-				Name:    "personal",
-				Aliases: []string{"p"},
-				Usage:   "run a personal task",
-				Action:  runPersonal,
-				Flags: []cli.Flag{
-					apiKeyFlag,
-					envFlag,
-					publishIntervalFlag,
-					numSubscriptionsFlag,
-					msgDataLengthFlag,
-					sseSubscriberFlag,
-				},
-			},
-			{
-				Name:    "sharded",
-				Aliases: []string{"s"},
-				Usage:   "run a sharded task",
-				Action:  runSharded,
-				Flags: []cli.Flag{
-					apiKeyFlag,
-					envFlag,
-					numChannelsFlag,
-					publishIntervalFlag,
-					msgDataLengthFlag,
-					numSubscriptionsFlag,
-					publisherFlag,
-				},
-			},
 			{
 				Name:    "composite",
 				Aliases: []string{"c"},
