+++ tasks/ably/ably.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"math/rand"
 	"strconv"
 	"sync"
+	"time"
 
 	"github.com/ably-forks/boomer"
 	"github.com/ably/ably-go/ably"
@@ -18,6 +20,79 @@ func newAblyClient(apiKey, env string) (*ably.RealtimeClient, error) {
 	options.Environment = env
 
 	return ably.NewRealtimeClient(options)
+}
+
+func millisecondTimestamp() int64 {
+	nanos := time.Now().UnixNano()
+	millis := nanos / int64(time.Millisecond)
+	return millis
+}
+
+func init() {
+	rand.Seed(time.Now().UnixNano())
+}
+
+var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+
+func randomString(length int) string {
+	b := make([]rune, length)
+	for i := range b {
+		b[i] = letters[rand.Intn(len(letters))]
+	}
+	return string(b)
+}
+
+func randomDelay(log log15.Logger) {
+	r := rand.Intn(60)
+	log.Info("introducing random delay", "seconds", r)
+	time.Sleep(time.Duration(r) * time.Second)
+	log.Info("continuing after random delay")
+}
+
+func publishOnInterval(
+	ctx context.Context,
+	publishInterval,
+	msgDataLength int,
+	channel *ably.RealtimeChannel,
+	delay int,
+	errorChannel chan<- error,
+	wg *sync.WaitGroup,
+	log log15.Logger,
+) {
+	log = log.New("channel", channel.Name)
+	log.Info("creating publisher", "period", publishInterval)
+
+	log.Info("introducing random delay before starting to publish", "seconds", delay)
+	time.Sleep(time.Duration(delay) * time.Second)
+	log.Info("continuing after random delay")
+
+	ticker := time.NewTicker(time.Duration(publishInterval) * time.Second)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ticker.C:
+			data := randomString(msgDataLength)
+			timePublished := strconv.FormatInt(millisecondTimestamp(), 10)
+
+			log.Info("publishing message", "size", len(data))
+			_, err := channel.Publish(timePublished, data)
+			if err != nil {
+				log.Error("error publishing message", "err", err)
+				boomer.RecordFailure("ably", "publish", 0, err.Error())
+				errorChannel <- err
+				ticker.Stop()
+				wg.Done()
+				return
+			}
+
+			boomer.RecordSuccess("ably", "publish", 0, 0)
+		case <-ctx.Done():
+			ticker.Stop()
+			wg.Done()
+			return
+		}
+	}
 }
 
 func reportSubscriptionToLocust(
